+++ compositor/meson.build
@@ -1,0 +1,6 @@
+executable(
+  meson.project_name(),
+  'main.cc',
+  install : true,
+  dependencies : libz11_public_dep,
+)
