+++ doc/CONTRIBUTING.ja.adoc
@@ -6,21 +6,9 @@ Z11はLinuxで動作するので、開発環境もLinuxを用います.
 
 ソースコードのビルドには link:https://mesonbuild.com/index.html[meson] を用います.
 
-=== Clone
-
-submoduleを含むので
-....
-git clone --recursive git@github.com:gray-armor/z11.git
-....
-
-普通にcloneしてしまったら、
-....
-git submodule update --init --recursive
-....
-
 === Custom Wayland
 
-カスタムされたWaylandを使うので、https://github.com/gray-armor/wayland をインストールする。
+カスタマイズされたWaylandを使うので、https://github.com/gray-armor/wayland をインストールする。
 
 ....
 $ git clone git@github.com:gray-armor/wayland.git
@@ -30,21 +18,30 @@ $ ninja -C build install
 
 === Build
 
+このレポジトリをビルド
+
 ....
-meson --prefix=$(pwd)/target build
-ninja -C build install
+$ git clone git@github.com:gray-armor/z11.git
+$ meson build
+$ ninja -C build install
 ....
 
 === Run
 
+新しくlibz11.soを作ったので、shared libraryの依存関係情報を更新する. (初回のみ)
+
+....
+$ ldconfig
+....
+
 run z server
 ....
-XDG_RUNTIME_DIR=~/.xdg ./target/bin/z11
+$ XDG_RUNTIME_DIR=~/.xdg z11
 ....
 
 run z client
 ....
-XDG_RUNTIME_DIR=~/.xdg ./target/bin/z11-simple-box
+$ XDG_RUNTIME_DIR=~/.xdg z11-simple-box
 ....
 
 === Tip
@@ -52,7 +49,7 @@ XDG_RUNTIME_DIR=~/.xdg ./target/bin/z11-simple-box
 Waylandで用いている通信内容をデバッグログとして出力することができます。
 `WAYLAND_DEBUG` 環境変数に `1` を設定してください。
 ....
-WAYLAND_DEBUG=1 XDG_RUNTIME_DIR=~/.xdg ./target/bin/z11
+$ WAYLAND_DEBUG=1 XDG_RUNTIME_DIR=~/.xdg z11
 ....
 
 === Coding Style
