+++ include/meson.build
@@ -1,0 +1,1 @@
+install_headers('libz11.h')
